Buttons and localStorage now working well. CSS not yet worked on but getting there I guess.
Also can now pass the prop "id" as a requirement for unique storage. Else, it would default to null which makes all null ids use the same storage which will be reflected on reload.

Next work is to make the interface buttons for the graph, passing of more props for configuration, and general fixes on styling too.